(INST) v.0.00.61 - Normalización CORE - STS
Normalización del SW STS.
Desarrollo de consultas STS.

+++ fecon/src/main/java/es/ldrsoftware/core/arq/BaseBS.java
@@ -87,7 +87,7 @@ public abstract class BaseBS extends BaseNotifyManager {
 		}
 	}
 	
-	protected void validateIntRange(long i, int min, int max, String notf) throws Exception {
+	protected void validateIntRange(long i, long min, long max, String notf) throws Exception {
 		if (i < min || i > max) {
 			notify(notf);
 		}
